Trait peano::ToInt

pub trait ToInt: Peano {
    fn to_int() -> i32;
}

Convert a Peano number to the integer it represents.

Once Rust implements associated constants, an associated constant will replace the function to_int().

Example:

use peano::{P2, ToInt};

assert_eq!(2, <P2 as ToInt>::to_int());

Required Methods

fn to_int() -> i32